+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# init release 0.2.0
+# init release 0.1.0
 - parse kv3
 - serde support
 - minor tests
+
+# 0.1.1
+- fix array handling with trailing commas
+- fixes some vphys parsing